fixed bug

+++ app.py
@@ -22,28 +22,31 @@ def final():
     cordinates = []
     data = request.values.get('arr_data')
     data = data.split(",")
-    for i, d in enumerate(data):
-        if i < len(data)-1:
-            lat = float(d[1:8])
-            lon = float(d[9:16])
-            cordinates.append([lat, lon])
-    
-    writeCsv('static/data.csv', cordinates)
-             
-    popSize = int(request.form["popSize"])
-    elitism = int(request.form["elitism"])
-    mutationRate = float(request.form["mutationRate"])
-    generations = int(request.form["generations"])
-    for i in execute_genetic(popSize=popSize, eliteSize=elitism, mutationRate=mutationRate, generations=generations):
-        bestRoute = i[0]
-        distance = i[1]
-        distance_list.append(distance)
-        map.drawLines(bestRoute)
-    print(distance)
-    
-    plot_distance_with_iterations(distance_list)
-    os.remove("static/data.csv")
-    return render_template('map.html', distance=distance)
+    if len(data) > 1:
+        for i, d in enumerate(data):
+            if i < len(data)-1:
+                lat = float(d[1:8])
+                lon = float(d[9:16])
+                cordinates.append([lat, lon])
+        
+        writeCsv('static/data.csv', cordinates)
+                
+        popSize = int(request.form["popSize"])
+        elitism = int(request.form["elitism"])
+        mutationRate = float(request.form["mutationRate"])
+        generations = int(request.form["generations"])
+        for i in execute_genetic(popSize=popSize, eliteSize=elitism, mutationRate=mutationRate, generations=generations):
+            bestRoute = i[0]
+            distance = i[1]
+            distance_list.append(distance)
+            map.drawLines(bestRoute)
+        print(distance)
+        
+        plot_distance_with_iterations(distance_list)
+        os.remove("static/data.csv")
+        return render_template('map.html')
+    else:
+        return render_template('map.html')
 
     return jsonify({'status':'200'})
 if __name__ == "__main__":
